florivo/urls.py
from django.contrib import admin
from django.urls import include, path
from debug_toolbar.toolbar import debug_toolbar_urls
from rest_framework import permissions
from drf_yasg.views import get_schema_view
from drf_yasg import openapi

from florivo.views import api_root_view

schema_view = get_schema_view(
   openapi.Info(
      title="Florivo - Flower Selling Website API",
      default_version='v1',
      description="API documentation for a flower selling web project",
      terms_of_service="https://www.google.com/policies/terms/",
      contact=openapi.Contact(email="contact@florivo.com"),
      license=openapi.License(name="BSD License"),
   ),
   public=True,
   permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    path("admin/", admin.site.urls),
    path('', api_root_view),
    path('api/v1/', include('api.urls'), name='api-root'),
    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
    path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
] + debug_toolbar_urls()
